+++ tomcat/src/main/java/org/apache/tomcat/util/http/HttpStatus.java
@@ -1,4 +1,4 @@
-package org.apache.catalina.connector.http;
+package org.apache.tomcat.util.http;
 
 public enum HttpStatus {
     OK(200),
